Creates ConfigurationContext and withConfiguration HOC
